+++ crates/module-system/module-implementations/sov-evm/src/call.rs
@@ -27,7 +27,9 @@ use crate::execution_config::EVM_EXECUTION_CONFIG;
 use crate::executor::{get_cfg_env, transact};
 #[cfg(feature = "native")]
 use crate::metrics::EvmTxMetrics;
-use crate::{gas_metering_mode, Evm, EvmRuntimeConfig, GasMeteringMode, PendingTransaction};
+use crate::{
+    gas_metering_mode, Evm, EvmChainSpec, EvmRuntimeConfig, GasMeteringMode, PendingTransaction,
+};
 use anyhow::{bail, Context as _};
 
 /// EVM call message.
@@ -71,11 +73,11 @@ where
         // Inside the EVM, we use nonces only for the CREATE operation.
         // The uniqueness check was performed before the call was dispatched.
         let account_nonce = self.get_account_nonce(signer, state)?;
-        let gas_limit = self.gas_limit(state);
-        let tx_env = create_tx_env(&tx, signer, account_nonce, gas_limit);
-        let tx = TxSignedAndRecovered::new(signer, tx, block_env.number.to::<u64>());
         let cfg = self.cfg(state)?;
         let cfg_env = get_cfg_env(&block_env, &cfg, None);
+        let gas_limit = self.gas_limit(state, &cfg.chain_spec);
+        let tx_env = create_tx_env(&tx, signer, account_nonce, gas_limit);
+        let tx = TxSignedAndRecovered::new(signer, tx, block_env.number.to::<u64>());
 
         Ok((cfg, cfg_env, block_env, tx_env, tx, pending_len))
     }
@@ -202,21 +204,22 @@ where
         Ok(())
     }
 
-    fn gas_limit(&self, state: &mut impl TxState<S>) -> u64 {
+    fn gas_limit(&self, state: &mut impl TxState<S>, spec: &EvmChainSpec) -> u64 {
         let gas_meter = state
             .try_as_basic_gas_meter()
             .expect("TxState should have BasicGasMeter");
         let funds = gas_meter.remaining_funds.map(|funds| funds.0).unwrap_or(0);
         let gas = gas_meter.remaining_gas.as_ref()[0];
         let price = gas_meter.gas_price.as_ref()[0].0;
-        match (funds, gas) {
+        let gas_limit = match (funds, gas) {
             (0, 0) => 0,
             (_, 0) => u64::MAX,
             (funds, gas) => {
                 let gas_from_funds = (funds / price).min(u64::MAX as u128) as u64;
                 gas.min(gas_from_funds)
             }
-        }
+        };
+        gas_limit.min(spec.tx_gas_limit.unwrap_or(u64::MAX))
     }
 
     fn sequencer_gas_used(&self, state: &mut impl TxState<S>) -> u64 {