+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2026-01-01
+- #2256 Enforce the EVM block gas limit and introduce the EVM tx gas limit in config.
 - #2279 Propagates fee information to ethereum block responses
 
 # 2025-12-22